Split service layer
